Use SFSafariViewController, fix UI on iPad
The UI didn't work at all on iPad for some reason, and after some
attempts to fix it the views in the split view started cropping
themselves slightly on the iPad and nothing else.  This was a dead end
for me and I chose to rewrite what was expressed in Storyboard in Swift.

I don't see myself using Storyboard ever again.