Reduce processing during prefetch
From the documentation for tableView(_:prefetchRowsAt:)

> ...the last index path corresponds to the row furthest from the
> visible area.

This means as the scrollbar goes down, the rows are in ascending order.
As it goes up, the rows are in descending order.  We only care about the
former, where we need to know if the highest row index is outside the
range os stories.

It strikes me that prefetch is currently spamming the server because
prefetch is called very often and a pending page doesn't claim a range
of rows.  In other words, every call to prefetch while is a page is
being loaded requests the next page.