Avoid server abuse during table prefetching
This relies on knowing about how many new stories will be loaded per
page (again hard coded as the server doesn't tell us ahead of time), the
number of pending pages, and what the highest row index marked for
prefetching was.  We need to check if the another page is necessary
after each load because the user might be forever waiting on a blank
screen otherwise.